requirements.txt
livekit-agents[google]~=1.2
python-dotenv
aiohttp
aiohttp-cors
